% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_exp_stat.R
\name{daily_exp_stat}
\alias{daily_exp_stat}
\title{Title}
\usage{
daily_exp_stat(DataExp, tE = 10, tM = 18)
}
\arguments{
\item{DataExp}{a dataframe obtained by the use of function `exposure_dat`}

\item{tE}{upper limit hour for evening exposure calculation, in a referential with 12h as origin (i.e 10 correspond to 22h)}

\item{tM}{lower limit hour for morning exposure calculation, in a referential with 12h as origin (i.e 18 correspond to 6h)}
}
\description{
Title
}
\examples{
Count_HB <- HB_to_counts(ODK_HB_R)
Count_HB$Enq <- 1
Data_Entomo <- Entomo_PHP_to_counts(Entomo_PHP)
Exposure <- exposure_dat(Count_HB, Data_Entomo, p = 0.92)
sumExp <- daily_exp_stat(Exposure, tE = 10, tM = 18)

}
