% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Entomo_PHP_to_counts.R
\name{Entomo_PHP_to_counts}
\alias{Entomo_PHP_to_counts}
\title{convert data from individual mosquitoes collection dataframe to counts of individuals per collection sites (giving 8 collections sites per place, 4 indoors, 4 outdoors)}
\usage{
Entomo_PHP_to_counts(data, v_hours, ndays = 1, genre = "Anopheles")
}
\arguments{
\item{data}{a data frame of individual mosquitoes with the following column :
`NumMission` survey of collection
`CodeVillage` village code (factor)
`PointDeCapture` collection site (i.e place where collection have been performed both indoors and outdoors) value : 1 to 4
`PosteDeCapture` collection point (indoors or oudoors), value = int or ext
`HeureDeCapture` beginning of the collection hours : passible values : 0 to 23
`Genre` genus of the individual mosquito (default "Anopheles")}

\item{v_hours}{a vector of sarting hours of collection i.e c(15:23,0:11) for collections lasting from 15h in the afternoon to 12h in the morning}

\item{ndays}{duration (in days) of collection during each survey in each collection point (default = 1)}

\item{genre}{the value (string) of the genus for which counts have to be computed (from factors in the `Genre` column, default : "Anopheles")}
}
\value{
a dataframe with hourly counts of mosquitoes in each collection point ("O1" "O2""O3""O4" "I1""I2" "I3" "I4"), counts summed indoor and outdoors ("No""Ni"),
 duration of collection indoors and outdoors (i.e., th number of collection point, d = 4), time in a new referential (t). Per villages and survey.
}
\description{
convert data from individual mosquitoes collection dataframe to counts of individuals per collection sites (giving 8 collections sites per place, 4 indoors, 4 outdoors)
}
\examples{
v_hours <- c(15:23,0:11)
ndays <- 2
Entomo_PHP_to_counts(Entomo_PHP, v_hours, ndays)
}
