% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_dat.R
\name{exposure_dat}
\alias{exposure_dat}
\title{Calculate hourly exposure to bite indices based on human and vector behaviors data
the generated dataframe can be used with functions `plot_exposure`, ` plot_behaviors` and `plot_behavior_H`for plotting purpose
and with `daily_exp_stat` and `simul_IC_exp` to calculate daily statistics of exposure}
\usage{
exposure_dat(Count_HB, Count_ento, p = 0.92)
}
\arguments{
\item{Count_HB}{a dataframe resulting from function `HB_to_counts` with an additional column specifying the survey number ("Enq", integer)}

\item{p}{the personal protection provided by an LLIN (default = 0.92 for Permanet 2 according to Corbel et al. 2010)}

\item{COunt_ento}{a dataframe resulting from function `Entomo_PHP_to_counts`}
}
\value{
a dataframe joining Count_HB and COunt_ento plus the following calculmated fields:
`pui`` proportion of LLIN user being indoor (rounded to deal with small neagtive value of U_outdoors)
`puo` proportion of LLIN user being outdoor
`pup` proportion of LLIN user being to bed and protected by a LLIN
`eui` user exposuExposure indoor
`euo` user exposuExposure outdoor
`eun` user exposuExposure under an LLIN
`eup` user exposure prevented by LLIN
}
\description{
Calculate hourly exposure to bite indices based on human and vector behaviors data
the generated dataframe can be used with functions `plot_exposure`, ` plot_behaviors` and `plot_behavior_H`for plotting purpose
and with `daily_exp_stat` and `simul_IC_exp` to calculate daily statistics of exposure
}
\examples{
Count_HB <- HB_to_counts(ODK_HB_R)
Count_HB$Enq <- 3
v_hours <- c(15:23,0:11)
Data_Entomo <- Entomo_PHP_to_counts(Entomo_PHP, v_hours)
exposure_dat(Count_HB, Data_Entomo, p = 0.92)

}
