% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HB_to_counts.R
\name{HB_to_counts}
\alias{HB_to_counts}
\title{Execute function `summarise_HB` on subsets (age classes and village) of a dataframe with data from interviews of 1 survey of human behavior (see MOIROUX et al. 2014)
Also weight count for all age classes according to household census to compute predicted count in the whole population
Also counts number of user outdoors and user indoors but not under net}
\usage{
HB_to_counts(data = data)
}
\arguments{
\item{data}{a data frame of individual behavior data with the following fields (order doesnt matter):
`rcpcodevillage`    village code (factor)
`rcpcodemenage`    household code
`rcpnbrenfant1`     nb of children (0-6y) in the household
`rcpnbrenfant2`    nb of children (6-18y) in the household
`rcpnbradulte`            nb of adults in the household
`rcptrcheage` individual age classes (factor)
`sexe` sexe of the individual (f or m)
`age` age in years of the individual
`dormirssmoust`, did the individual used a mosquito net ? (value = "oui" or "non"),
`hintmaison`, the hour the individual went indoors in the evening, in HH:MM:SS character format (for ex. "20:30:00")
`hsortiemaison`, the hour the individual went outdoors in the morning
`hcoucher`, the hour the individual went to sleep
`hlever`, the hour the individual wake up}
}
\value{
a data frame of counts of individuals in various compartment per hours according to the age classes and the village (to be used for bite exposure calculation)
}
\description{
`HB_to_counts` returns hourly counts of individual humans in each compartiments (indoor, outdoors, under nets) from a data frame specifying (see MOIROUX et al. 2014)
}
\examples{
HB_to_counts(ODK_HB_R)
}
