% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_behavior_H}
\alias{plot_behavior_H}
\title{multi panel plot (per age classes) of human behavior}
\usage{
plot_behavior_H(Exposure, vil = c("DOG"), enq = 1, nc = 6)
}
\arguments{
\item{nc}{}
}
\description{
multi panel plot (per age classes) of human behavior
}
