% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_behaviors}
\alias{plot_behaviors}
\title{plot human and vector behavior on the same graph (multi-panel plot, one plot per survey)}
\usage{
plot_behaviors(Exposure, vil, age = "pop")
}
\arguments{
\item{age}{}
}
\description{
plot human and vector behavior on the same graph (multi-panel plot, one plot per survey)
}
