% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_exposure}
\alias{plot_exposure}
\title{making multi panel plot of exposure to bite (one panel per entomological survey for one village and one age classes)}
\usage{
plot_exposure(Exposure, vil, age = "pop")
}
\arguments{
\item{Exposure}{a dataframe obtained by the use of function `exposure_dat`}

\item{vil}{one village code (string) from Exposure$Vil}

\item{age}{one age classes (string) from Exposure$Age}
}
\value{
a ggplot object
}
\description{
making multi panel plot of exposure to bite (one panel per entomological survey for one village and one age classes)
}
\examples{
Exposure <- exposure_dat(Count_HB, Data_Entomo, p = 0.92)
plot_exposure(Exposure, vil= "DOG", age = "pop")

}
