% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul_IC_exp.R
\name{simul_IC_exp}
\alias{simul_IC_exp}
\title{Calculates exposure indices and simulates confidence intervals based on bootstraping methods}
\usage{
simul_IC_exp(DataExp, p = 0.92, ndays = 1, nsim = 5000, tE = 10,
  tM = 18)
}
\arguments{
\item{DataExp}{a dataframe obtained by the use of function `exposure_dat`}

\item{p}{the personal protection provided by an LLIN (default = 0.92 for Permanet 2 according to Corbel et al. 2010)}

\item{ndays}{duration (in days) of collection during each survey (default = 1)}

\item{nsim}{number of simulations (default = 5000)}

\item{tE}{upper limit hour for evening exposure calculation, in a referential with 12h as origin (i.e 10 correspond to 22h)}

\item{tM}{lower limit hour for morning exposure calculation, in a referential with 12h as origin (i.e 18 correspond to 6h)}
}
\value{
a dataframe with 7 columns :
"Vil"     village code (factor)
"Enq"     survey number (integer)
"Age"     age classes (factor)
"var_exp" name of the calculated exposure variable (Eff, Peui, Peuni, PeuE, PeneE, PeuM, PenuM) (factor)
"mean"    mean value (real)
"lo95"    lower bound of 95% CI of mean  (real)
"hi95"   higer bound of 95% CI of mean  (real)
}
\description{
Calculates exposure indices and simulates confidence intervals based on bootstraping methods
}
