% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_HB.R
\name{summarise_HB}
\alias{summarise_HB}
\title{Count individuals in different states (indoors, outdoors, sleeping) for both user and non-users of nets at each hour (from 0h to 23h)}
\usage{
summarise_HB(data)
}
\arguments{
\item{data}{a data frame of individual behavior data with the following fields:
`dormirssmoust`, did the individual used a mosquito net ? (value = "oui" or "non"),
`hintmaison`, the hour the individual went indoors in the evening, in HH:MM:SS character format (for ex. "20:30:00")
`hsortiemaison`, the hour the individual went outdoors in the morning
`hcoucher`, the hour the individual went to sleep
`hlever`, the hour the individual wake up}
}
\value{
a dataframe with 6 columns of counts (N_Ind, N_User, N_In, N_Sleep, N_Mild, N_User_In) and 24 lines (one per hours)
}
\description{
`summarise_HB` returns hourly counts of individual humans in 3 compartiments (indoor, outdoors, under nets) from individual data (hours) or behavior.
Used in `HB_to_counts` function
}
\examples{
subset_hb <- subset(ODK_HB_R,rcpcodevillage == "DOG")
summarise_HB(subset_hb)
}
