% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_exposure}
\alias{plot_exposure}
\title{makes multi panel plot of exposure to bite (one panel per entomological survey for one village and one age classes)}
\usage{
plot_exposure(Exposure, vil, age = "pop")
}
\arguments{
\item{Exposure}{a dataframe obtained by the use of function `exposure_dat`}

\item{vil}{one village code (string) from Exposure$Vil}

\item{age}{one age classes (string) from Exposure$Age}
}
\value{
a ggplot object
}
\description{
makes multi panel plot of exposure to bite (one panel per entomological survey for one village and one age classes)
}
\examples{
Count_HB <- HB_to_counts(ODK_HB_R)
Count_HB$Enq <- 3
v_hours <- c(15:23,0:11)
Data_Entomo <- Entomo_PHP_to_counts(Entomo_PHP, v_hours)
Exposure <- exposure_dat(Count_HB, Data_Entomo, p = 0.92)
plot_exposure(Exposure, vil= "DOG", age = "pop")

}
